requirements-test.txt
# Testing requirements for Ultraloq Wifi integration

# Core testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Environment variable support
python-dotenv>=1.0.0

# HTTP client (same as used in integration)
aiohttp>=3.8.0

# Optional: Code coverage
pytest-cov>=4.0.0

# Optional: Test reporting
pytest-html>=3.1.0